# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tasty Crousty - Smart food ordering and delivery platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: restaurantId
          in: query
          schema:
            type: ["integer", "null"]
        - name: driverId
          in: query
          schema:
            type: ["integer", "null"]
        - name: cityId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{orderId}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderBody"
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/start-preparing:
    post:
      operationId: startPreparing
      tags: [orders]
      summary: Restaurant marks order as preparing
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order preparing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{orderId}/status-history:
    get:
      operationId: getOrderStatusHistory
      tags: [orders]
      summary: Get order status history
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderStatusHistory"
  /orders/{orderId}/qr:
    get:
      operationId: getOrderQR
      tags: [orders]
      summary: Get QR delivery token for order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QRToken"

  # DISPATCH
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
          enum:
            - draft
            - pending_dispatch
            - dispatching_driver
            - driver_assigned
            - awaiting_customer_confirmation
            - needs_update
            - confirmation_failed
            - confirmed_for_preparation
            - preparing
            - ready_for_pickup
            - picked_up
            - on_the_way
            - arriving_soon
            - delivered
            - cancelled
            - failed
            - refunded
        customerId:
          type: integer
        restaurantId:
          type: integer
        restaurantName:
          type: string
        driverId:
          type: ["integer", "null"]
        driverName:
          type: ["string", "null"]
        deliveryAddress:
          type: string
        deliveryLandmark:
          type: ["string", "null"]
        deliveryFloor:
          type: ["string", "null"]
        deliveryInstructions:
          type: ["string", "null"]
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash_on_delivery, online]
        paymentStatus:
          type: string
          enum: [pending, authorized, paid, failed, refunded, cash_on_delivery]
        estimatedDeliveryMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - status
        - customerId
        - restaurantId
        - restaurantName
        - deliveryAddress
        - subtotal
        - deliveryFee
        - total
        - paymentMethod
        - paymentStatus
        - createdAt
        - updatedAt
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/OrderStatusHistory"
            qrToken:
              type: ["string", "null"]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - id
        - productId
        - productName
        - quantity
        - price
    OrderStatusHistory:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        status:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
        createdBy:
          type: ["string", "null"]
      required:
        - id
        - orderId
        - status
        - createdAt
    PaginatedOrders:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - limit
        - totalPages
    CreateOrderBody:
      type: object
      properties:
        restaurantId:
          type: integer
        deliveryAddress:
          type: string
        deliveryLandmark:
          type: ["string", "null"]
        deliveryFloor:
          type: ["string", "null"]
        deliveryInstructions:
          type: ["string", "null"]
        deliveryPhone:
          type: ["string", "null"]
        zoneId:
          type: ["integer", "null"]
        paymentMethod:
          type: string
          enum: [cash_on_delivery, online]
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              notes:
                type: ["string", "null"]
            required:
              - productId
              - quantity
      required:
        - restaurantId
        - deliveryAddress
        - paymentMethod
        - items
    CancelOrderBody:
      type: object
      properties:
        reason:
          type: ["string", "null"]
    QRToken:
      type: object
      properties:
        token:
          type: string
        orderId:
          type: integer
        expiresAt:
          type: ["string", "null"]
      required:
        - token
        - orderId

    # DISPATCH
...